[fmt] %x is for unsigned, prevent UB in reading to plain int
According to HTTP 1.1 spec the chunk size is transferred as a
hexadecimal number followed by \r\n as a line separator, followed by
a chunk of data of the given size. As gHttp assumes chunk size of
int type, i just read size to unsigned as needed and check actual
bounds.